+++ app/src/main/kotlin/paufregi/garminfeed/presentation/quickedit/QuickEditState.kt
@@ -1,10 +1,10 @@
 package paufregi.garminfeed.presentation.quickedit
 
 import paufregi.garminfeed.core.models.Activity
-import paufregi.garminfeed.core.models.ActivityProfile
+import paufregi.garminfeed.core.models.Profile
 
 data class QuickEditState(
     val loading: Boolean = false,
     val activities: List<Activity> = emptyList(),
-    val profiles: List<ActivityProfile> = emptyList()
+    val profiles: List<Profile> = emptyList()
 )
